+++ variant/validators/genomic_delins.py
@@ -125,13 +125,13 @@ class GenomicDelIns(DelInsBase):
                         self.get_hgvs_expr(classification, t, s, False)
                     allele = self.get_allele_from_hgvs(hgvs_expr, errors)
 
-                if allele:
-                    mane_transcripts_dict[hgvs_expr] = {
-                        'classification_token': s,
-                        'transcript_token': t,
-                        'is_ensembl_transcript': is_ensembl_transcript
-                    }
+                mane_transcripts_dict[hgvs_expr] = {
+                    'classification_token': s,
+                    'transcript_token': t,
+                    'is_ensembl_transcript': is_ensembl_transcript
+                }
 
+                if allele:
                     len_of_seq = self.seqrepo_access.len_of_sequence(t)
                     is_len_lt_end = len_of_seq < int(s.end_pos_del) - 1
                     is_len_lt_start = \
